pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "agentic-ai-security"
version = "0.1.0"
description = "AI-powered source code security analyzer with MCP integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Krishcalin"}]
keywords = ["security", "scanner", "mcp", "ai", "sast", "prompt-injection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "structlog>=23.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
llm = ["anthropic>=0.30"]
full = ["anthropic>=0.30", "tree-sitter>=0.20", "python-Levenshtein>=0.21"]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.1.0", "mypy>=1.5"]

[project.scripts]
agentic-scan = "cli.main:cli"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
